refactor(linter): Move github linters to new file
