+++ tasks/linter/github.py
@@ -1,0 +1,36 @@
+"""GitHub-related linting tasks (PRs and GitHub Actions)"""
+
+import os
+import sys
+
+from invoke.exceptions import Exit
+from invoke.tasks import task
+
+from tasks.libs.ciproviders.github_api import GithubAPI
+from tasks.libs.common.color import color_message
+from tasks.libs.common.constants import GITHUB_REPO_NAME
+from tasks.libs.common.utils import is_pr_context
+
+
+@task
+def releasenote(ctx):
+    """Lints release notes with Reno."""
+
+    branch = os.environ.get("BRANCH_NAME")
+    pr_id = os.environ.get("PR_ID")
+
+    run_check = is_pr_context(branch, pr_id, "release note")
+    if run_check:
+        github = GithubAPI(repository=GITHUB_REPO_NAME, public_repo=True)
+        if github.is_release_note_needed(pr_id):
+            if not github.contains_release_note(pr_id):
+                print(
+                    f"{color_message('Error', 'red')}: No releasenote was found for this PR. Please add one using 'reno'"
+                    ", see https://datadoghq.dev/datadog-agent/guidelines/contributing/#reno"
+                    ", or apply the label 'changelog/no-changelog' to the PR.",
+                    file=sys.stderr,
+                )
+                raise Exit(code=1)
+            ctx.run("reno lint")
+        else:
+            print("'changelog/no-changelog' label found on the PR: skipping linting")
